+++ packages/loading-scripts/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.0.7](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.6...@rudderstack/analytics-js-loading-scripts@3.0.7) (2024-05-24)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.0.6`
 ## [3.0.6](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.5...@rudderstack/analytics-js-loading-scripts@3.0.6) (2024-05-10)
 
 ### Dependency Updates